+++ judge/codeforces/900_rated/29.cpp
@@ -18,7 +18,7 @@ void solve()
     int s;
     cin >> s;
 
-    ll ans = 0;
+    int ans = 0;
 
     while (true)
     {
